Actualizadas clases con herencia
